fix(ci,ui): 修掉 6 条 CodeQL high 告警，补上发布目标的祖先校验
- validate-draft-release.sh 增加「target_sha 必须在 main 历史上」的校验。此前有 write 权限的人可以
  建一个指向任意 commit 的 draft release 再 dispatch，那份代码就会被编译并签进正式安装包；唯一的
  祖先检查跑在 checkout 之后的不可信工作树里，而负责编译的 TSF job 根本没有这道检查。校验放在
  prepare（默认分支上下文）里，所有构建 job 都 needs 它。
- 字体名引用只转义了 `"` 没转义反斜线，以 `\` 结尾的字体名能吃掉闭合引号让后续内容按 CSS 解析。
  两处（appearance.ts、shared.ts）都改为先转义转义符。
- voice_input_service.cpp: float 平方先提升到 double
- engine-bump-triage.yml 顶层权限从 write 收到 read，写权限下放到 job；runner 与同仓其他 workflow
  统一为 ubuntu-24.04

新增 CONTRIBUTING.md，把散在各 AGENTS.md / 组件 README 里的环境要求、构建顺序、打包步骤和提交前
检查串成外部贡献者能照着走完的一条路径，并从 README 首屏链出。

+++ server/src/voice-input/voice_input_service.cpp
@@ -756,7 +756,9 @@ void AudioCallback(const float *samples, std::size_t frames)
 {
     if (!samples || !g_recording) return;
     double sum = 0.0;
-    for (std::size_t i = 0; i < frames; ++i) sum += samples[i] * samples[i];
+    // Widen before squaring. float * float is evaluated at float precision and only then converted,
+    // which throws away the accuracy the double accumulator exists to keep.
+    for (std::size_t i = 0; i < frames; ++i) sum += static_cast<double>(samples[i]) * static_cast<double>(samples[i]);
     const float rms = frames ? static_cast<float>(std::sqrt(sum / frames)) : 0.0f;
     // Compress the visual dynamic range so quiet speech still produces a clear waveform,
     // while keeping low-level room/microphone noise close to rest.
